requirements.txt
pyyaml>=6.0
requests>=2.31.0
google-genai>=1.0.0

